+++ src/components/SearchPageComponents/SelectSubmitIngredients/RecipeSearchButton.tsx
@@ -1,6 +1,6 @@
 import Button from '@mui/material/Button';
 import Icon from '@mui/material/Icon';
-import { RecipeSearchButtonProps } from '../../common/types';
+import { RecipeSearchButtonProps } from '../../../common/types';
 
 const RecipeSearchButton = (props: RecipeSearchButtonProps) => {
   const { isButtonDisabled, handleSearchForRecipes } = props;
