+++ src/scss/components/_select-prompt.scss
@@ -13,8 +13,7 @@
 @media only screen and (max-width: 768px) {
   .select-prompt {
     width: 100%;
-    height: 22%;
-
+    height: 21vh;
     padding-top: 0%;
 
     display: flex;
