+++ src/scss/components/_selected-ingredient-list.scss
@@ -1,6 +1,5 @@
 // desktop
 .selected-ingredient-list {
-  // padding-top: 50%;
   display: flex;
   flex-direction: column;
   height: 300px;
@@ -27,12 +26,11 @@
 //mobile
 @media only screen and (max-width: 768px) {
   .selected-ingredient-list {
-    padding-top: 50%;
-    height: 60%;
+    height: 60vh;
     width: 100%;
+    padding-top: 12vh;
     max-width: 100%;
     display: flex;
-    align-items: left;
 
     & > li {
       font-size: large;
